Remove the refactoring work and make the test passed
1. Remove all of refactoring work
2. make the test pass

+++ kombu/transport/filesystem.py
@@ -88,7 +88,6 @@ Transport Options
 * ``processed_folder`` - directory where are backed up processed files.
 """
 
-import glob
 import os
 import shutil
 import tempfile
@@ -106,10 +105,10 @@ from kombu.utils.objects import cached_property
 from . import virtual
 
 VERSION = (1, 0, 0)
-__version__ = ".".join(map(str, VERSION))
+__version__ = '.'.join(map(str, VERSION))
 
 # needs win32all to work on Windows
-if os.name == "nt":
+if os.name == 'nt':
 
     import pywintypes
     import win32con
@@ -124,15 +123,15 @@ if os.name == "nt":
     def lock(file, flags):
         """Create file lock."""
         hfile = win32file._get_osfhandle(file.fileno())
-        win32file.LockFileEx(hfile, flags, 0, 0xFFFF0000, __overlapped)
+        win32file.LockFileEx(hfile, flags, 0, 0xffff0000, __overlapped)
 
     def unlock(file):
         """Remove file lock."""
         hfile = win32file._get_osfhandle(file.fileno())
-        win32file.UnlockFileEx(hfile, 0, 0xFFFF0000, __overlapped)
+        win32file.UnlockFileEx(hfile, 0, 0xffff0000, __overlapped)
 
 
-elif os.name == "posix":
+elif os.name == 'posix':
 
     import fcntl
     from fcntl import LOCK_EX, LOCK_NB, LOCK_SH  # noqa
@@ -147,10 +146,12 @@ elif os.name == "posix":
 
 
 else:
-    raise RuntimeError("Filesystem plugin only defined for NT and POSIX platforms")
+    raise RuntimeError(
+        'Filesystem plugin only defined for NT and POSIX platforms')
 
 
-exchange_queue_t = namedtuple("exchange_queue_t", ["routing_key", "pattern", "queue"]) 
+exchange_queue_t = namedtuple("exchange_queue_t",
+                              ["routing_key", "pattern", "queue"])
 
 
 class Channel(virtual.Channel):
@@ -162,8 +163,10 @@ class Channel(virtual.Channel):
     def _get_exchange_file_obj(self, exchange, mode="rb"):
         filename = f"{exchange}.exchange"
         filename = os.path.join(self.control_folder, filename)
+        os.makedirs(self.control_folder, exist_ok=True)
+        f_obj = open(filename, mode)
+
         try:
-            f_obj = open(filename, mode)
             if "w" in mode:
                 lock(f_obj, LOCK_EX)
             yield f_obj
@@ -184,7 +187,8 @@ class Channel(virtual.Channel):
 
     def _queue_bind(self, exchange, routing_key, pattern, queue):
         queues = self.get_table(exchange)
-        queue_val = exchange_queue_t(routing_key or "", pattern or "", queue or "")
+        queue_val = exchange_queue_t(routing_key or "", pattern or "",
+                                     queue or "")
         if queue_val not in queues:
             queues.insert(0, queue_val)
         with self._get_exchange_file_obj(exchange, "wb") as f_obj:
@@ -196,27 +200,32 @@ class Channel(virtual.Channel):
 
     def _put(self, queue, payload, **kwargs):
         """Put `message` onto `queue`."""
-        filename = "{}_{}.{}.msg".format(
-            int(round(monotonic() * 1000)), uuid.uuid4(), queue
-        )
+        filename = '{}_{}.{}.msg'.format(int(round(monotonic() * 1000)),
+                                         uuid.uuid4(), queue)
         filename = os.path.join(self.data_folder_out, filename)
 
         try:
-            f = open(filename, "wb")
+            f = open(filename, 'wb')
             lock(f, LOCK_EX)
             f.write(str_to_bytes(dumps(payload)))
         except OSError:
-            raise ChannelError(f"Cannot add file {filename!r} to directory")
+            raise ChannelError(
+                f'Cannot add file {filename!r} to directory')
         finally:
             unlock(f)
             f.close()
 
     def _get(self, queue):
         """Get next message from `queue`."""
-        folder = self._ls_queue(queue)
+        queue_find = '.' + queue + '.msg'
+        folder = os.listdir(self.data_folder_in)
         folder = sorted(folder)
         while len(folder) > 0:
             filename = folder.pop(0)
+
+            # only handle message for the requested queue
+            if filename.find(queue_find) < 0:
+                continue
 
             if self.store_processed:
                 processed_folder = self.processed_folder
@@ -225,37 +234,44 @@ class Channel(virtual.Channel):
 
             try:
                 # move the file to the tmp/processed folder
-                shutil.move(filename, processed_folder)
+                shutil.move(os.path.join(self.data_folder_in, filename),
+                            processed_folder)
             except OSError:
                 pass  # file could be locked, or removed in meantime so ignore
 
             filename = os.path.join(processed_folder, filename)
             try:
-                f = open(filename, "rb")
+                f = open(filename, 'rb')
                 payload = f.read()
                 f.close()
                 if not self.store_processed:
                     os.remove(filename)
             except OSError:
-                raise ChannelError(f"Cannot read file {filename!r} from queue.")
+                raise ChannelError(
+                    f'Cannot read file {filename!r} from queue.')
 
             return loads(bytes_to_str(payload))
 
         raise Empty()
 
-    def _ls_queue(self, queue):
-        """List all messages in the `queue`"""
-        queue_pattern = os.path.join(self.data_folder_in, f"*.{queue}.msg")
-        return glob.glob(queue_pattern)
-
     def _purge(self, queue):
         """Remove all messages from `queue`."""
         count = 0
-        queue_messages = self._ls_queue(queue)
-        for filename in queue_messages:
+        queue_find = '.' + queue + '.msg'
+
+        folder = os.listdir(self.data_folder_in)
+        while len(folder) > 0:
+            filename = folder.pop()
             try:
+                # only purge messages for the requested queue
+                if filename.find(queue_find) < 0:
+                    continue
+
+                filename = os.path.join(self.data_folder_in, filename)
                 os.remove(filename)
+
                 count += 1
+
             except OSError:
                 # we simply ignore its existence, as it was probably
                 # processed by another worker
@@ -265,7 +281,20 @@ class Channel(virtual.Channel):
 
     def _size(self, queue):
         """Return the number of messages in `queue` as an :class:`int`."""
-        return self._ls_queue(queue)
+        count = 0
+
+        queue_find = f'.{queue}.msg'
+        folder = os.listdir(self.data_folder_in)
+        while len(folder) > 0:
+            filename = folder.pop()
+
+            # only handle message for the requested queue
+            if filename.find(queue_find) < 0:
+                continue
+
+            count += 1
+
+        return count
 
     @property
     def transport_options(self):
@@ -273,42 +302,43 @@ class Channel(virtual.Channel):
 
     @cached_property
     def data_folder_in(self):
-        return self.transport_options.get("data_folder_in", "data_in")
+        return self.transport_options.get('data_folder_in', 'data_in')
 
     @cached_property
     def data_folder_out(self):
-        return self.transport_options.get("data_folder_out", "data_out")
+        return self.transport_options.get('data_folder_out', 'data_out')
 
     @cached_property
     def store_processed(self):
-        return self.transport_options.get("store_processed", False)
+        return self.transport_options.get('store_processed', False)
 
     @cached_property
     def processed_folder(self):
-        return self.transport_options.get("processed_folder", "processed")
+        return self.transport_options.get('processed_folder', 'processed')
 
     @cached_property
     def control_folder(self):
-        return self.transport_options.get("control_folder", "control")
+        return self.transport_options.get('control_folder', 'control')
 
 
 class Transport(virtual.Transport):
     """Filesystem Transport."""
 
     implements = virtual.Transport.implements.extend(
-        asynchronous=False, exchange_type=frozenset(["direct", "topic", "fanout"])
+        asynchronous=False,
+        exchange_type=frozenset(['direct', 'topic', 'fanout'])
     )
 
     Channel = Channel
     # filesystem backend state is global.
     global_state = virtual.BrokerState()
     default_port = 0
-    driver_type = "filesystem"
-    driver_name = "filesystem"
+    driver_type = 'filesystem'
+    driver_name = 'filesystem'
 
     def __init__(self, client, **kwargs):
         super().__init__(client, **kwargs)
         self.state = self.global_state
 
     def driver_version(self):
-        return "N/A"
+        return 'N/A'
